requirements.txt
mysql-connector-python~=8.0.32
requests~=2.31.0
Flask~=2.2.3
Flask-Cors
bcrypt~=4.0.1
resend~=0.4.0
PyJWT~=2.7.0
Pillow~=9.4.0
replicate~=0.8.4
python-dotenv~=1.0.0